[breaking] fix EXPath Package dependency logic
Under 2.4.0, “4.0.0-SNAPSHOT” would satisfy the EXPath Package SemVer template semver-max=“3”, since “3” expands to <4.0.0 and 4.0.0-SNAPSHOT < 4.0.0. This commit disregards a version’s pre-release identifier when checking if an EXPath Package SemVer template is satisfied by a version. So “4.0.0-SNAPSHOT” is treated as “4.0.0” and thus fails the dependency check <4.0.0.

This commit also adds handling for EXPath Package dependency versioning attributes, “versions” and “semver”. And ensures that the versioning attributes are mutually exclusive according to the EXPath spec.

This commit also updates the EXPath Package SemVer templates regex to require major.minor and reject pre-release and build metadata.